requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
pydantic>=1.10.13
pydantic-settings>=2.6.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.26.0
aiohttp>=3.9.5
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
python-multipart>=0.0.9
